requirements.txt
anthropic==0.18.1
python-dotenv==1.0.1
flask==3.0.2
flask-cors==4.0.0
markdown==3.5.2
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.5
werkzeug==3.0.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
email-validator==2.1.0.post1 